Revert "Merge branch 'enable_spyfish_tut8' into dev"
This reverts commit 0cb588c8e06e0f7983d878eca84e2ed96bc1105c, reversing
changes made to c83f5dc2752101e43e0d9afb8d4c47961e092847.
