+++ kso_utils/server_utils.py
@@ -122,14 +122,14 @@ def download_init_csv(project: Project, init_keys: list, server_connection: dict
             db_initial_info[str("server_" + init_key + "_csv")] = server_csv
 
             # Specify the local path for the csv
-            local_i_csv = Path(project.csv_folder, Path(server_csv).name)
+            local_i_csv = str(Path(project.csv_folder, Path(server_csv).name))
 
             # Download the csv
             download_object_from_s3(
                 client=server_connection["client"],
                 bucket=project.bucket,
                 key=server_csv,
-                filename=str(local_i_csv),
+                filename=local_i_csv,
             )
 
     else:
